requirement.txt
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.42.0
humanize==4.7.0
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pygame==2.5.0
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
PyYAML==6.0.1
scipy==1.11.1
six==1.16.0
sympy==1.12
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
